Code improvements Moved "FormatValue" function from TDisassembler to own class TValueProcessor. Added tests for new TValueProcessor class. Fixed value presentation bug for binary number is Disassembler

+++ src/formula.cpp
@@ -61,7 +61,7 @@ int64_t TFormula::Evaluate(TString term)
     // Extract the left part of the formula
     TString left_term = term.SubString(0, operator_pos);
     // Extract the right part of the formula
-    TString right_term = term.SubString(operator_pos + 1);
+    TString right_term = term.SubString(static_cast<std::size_t>(operator_pos) + 1);
 
     // Recursively compute both parts of the formula
     const auto left_value = this->Evaluate(left_term);
